Add timestamp to URLs in debug mode.
Adds a URL parameter with a timestamp to the end of individual
JS and CSS URLs when set to debug mode. This is to avoid caching
in certain cases when developing locally, such as working on a 
test Facebook app.

Also fixes the tests to handle this change and adds media files
to the test settings that were previously empty so that the tests
actually test the right output (previously they were comparing two
empty strings due to the empty minify bundles).