Fix hard-coded MEDIA stuff
Django 1.4 uses STATIC/STATIC_URL for static assets and MEDIA/MEDIA_URL
for uploaded files and things like that.

This adds an additional setting which switches jingo-minify to use the
STATIC settings. Doing it this way maintains backwards compatibility
with existing Django sites using jingo-minify, but lets new Django 1.4
based sites use it and have things in the right place.