Use Django's static finder if appropriate.
When DEBUG is true and the new USE_STATIC setting is set, let Django see
if it can find the static file. This allows files to be found that are
in 'project/app/static', instead of only the centralized static
directory.

This change really only effects development, but it prevents having to
run './manage.py collecstatic' after every change while in development.
Instead files static files will be updated as they are changed so edit,
save, refresh works.