Make possible to define a custom static root
We don't want to look for statics in STATIC_ROOT, which is to be used by
django to collect statics.
And we neither want to look for statics in MEDIA_ROOT, indeed.
Also, we don't want to generate our minified/compressed statics in
STATIC_ROOT, as it's not taken into account by django serve method
when in DEBUG=True